# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PerioVision AI API",
    "version": "2.0",
    "description": "Research prototype for periodontal decision support. Every JSON response uses the envelope {data, meta, error, mode}. Authenticate with POST /api/auth/login, then send 'Authorization: Bearer <access_token>'. x-permission names the RBAC permission each operation requires (see docs/SECURITY.md)."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "AnalyzeIn": {
        "additionalProperties": false,
        "properties": {
          "upload_id": {
            "pattern": "^[a-f0-9]{32}$",
            "title": "Upload Id",
            "type": "string"
          },
          "patient_id": {
            "minimum": 1,
            "title": "Patient Id",
            "type": "integer"
          },
          "visit_date": {
            "anyOf": [
              {
                "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Visit Date"
          }
        },
        "required": [
          "upload_id",
          "patient_id"
        ],
        "title": "AnalyzeIn",
        "type": "object"
      },
      "Envelope": {
        "type": "object",
        "required": [
          "data",
          "meta",
          "error",
          "mode"
        ],
        "properties": {
          "data": {
            "description": "Payload (object, list or null on error)"
          },
          "meta": {
            "type": "object",
            "description": "Extra information such as counts"
          },
          "error": {
            "nullable": true,
            "type": "object",
            "properties": {
              "code": {
                "type": "integer"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            }
          },
          "mode": {
            "type": "string",
            "enum": [
              "demo",
              "live"
            ]
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/analyses": {
      "post": {
        "summary": "Run the full analysis pipeline on an upload",
        "operationId": "analysis_create_analysis_post",
        "tags": [
          "analysis"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          },
          "422": {
            "description": "Validation failed; error.details lists each field"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "analysis:run",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeIn"
              }
            }
          }
        }
      },
      "get": {
        "summary": "Recent analyses visible to you",
        "operationId": "analysis_list_analyses_get",
        "tags": [
          "analysis"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "analysis:read"
      }
    },
    "/api/analyses/{analysis_id}": {
      "get": {
        "summary": "Analysis detail: teeth, uncertainty, risk, review status",
        "operationId": "analysis_get_analysis_get",
        "tags": [
          "analysis"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "analysis:read",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/analyses/{analysis_id}/image/{layer}": {
      "get": {
        "summary": "Decrypted image layer (radiograph, annotated, gradcam)",
        "operationId": "analysis_analysis_image_get",
        "tags": [
          "analysis"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "analysis:read",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "layer",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/audit/anchor": {
      "post": {
        "summary": "Publish (anchor) a Merkle root now",
        "operationId": "audit_anchor_post",
        "tags": [
          "audit"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "audit:verify"
      }
    },
    "/api/audit/logs": {
      "get": {
        "summary": "Browse the tamper-evident audit log",
        "operationId": "audit_logs_get",
        "tags": [
          "audit"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "audit:read"
      }
    },
    "/api/audit/verify": {
      "get": {
        "summary": "Verify the hash chain and Merkle anchors",
        "operationId": "audit_verify_get",
        "tags": [
          "audit"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "audit:verify"
      }
    },
...
  }
}